@@ -30,69 +30,43 @@ class Plaid
 	const API_VERSION = "2020-09-14";
 
 	/**
-	 * Plaid client Id.
-	 *
-	 * @var string
-	 */
-	protected $client_id;
-
-	/**
-	 * Plaid client secret.
-	 *
-	 * @var string
-	 */
-	protected $client_secret;
-
-	/**
-	 * Plaid API host environment.
-	 *
-	 * @var string
-	 */
-	protected $environment = "production";
-
-	/**
 	 * Plaid API environments and matching hostname.
 	 *
 	 * @var array<string,string>
 	 */
-	protected $plaidEnvironments = [
+	protected array $plaidEnvironments = [
 		"production" => "https://production.plaid.com/",
 		"development" => "https://development.plaid.com/",
 		"sandbox" => "https://sandbox.plaid.com/",
 	];
 
 	/**
-	 * ClientInterface instance.
-	 *
-	 * @var ClientInterface|null
-	 */
-	protected $httpClient;
-
-	/**
 	 * Resource instance cache.
 	 *
 	 * @var array<AbstractResource>
 	 */
-	protected $resource_cache = [];
+	protected array $resource_cache = [];
+
 
 	/**
-	 * @param string $client_id
-	 * @param string $client_secret
-	 * @param string $environment Possible values are: production, development, sandbox
+	 * @param string $client_id Plaid client Id.
+	 * @param string $client_secret Plaid client secret.
+	 * @param string $environment Plaid API host environment. Possible values are: production, development, sandbox.
 	 * @throws UnexpectedValueException
 	 */
 	public function __construct(
-		string $client_id,
-		string $client_secret,
-		string $environment = "production")
+		protected string $client_id,
+		protected string $client_secret,
+		protected string $environment = "production",
+		protected ?ClientInterface $httpClient = null)
 	{
 		if( !\array_key_exists($environment, $this->plaidEnvironments) ){
 			throw new UnexpectedValueException("Invalid environment. Environment must be one of: production, development, or sandbox.");
 		}
 
-		$this->client_id = $client_id;
-		$this->client_secret = $client_secret;
-		$this->environment = $environment;
+		if( empty($this->httpClient) ){
+			$this->httpClient = new Shuttle;
+		}
 	}
 
 	/**
@@ -120,7 +94,7 @@ class Plaid
 			 * @var AbstractResource $resource_instance
 			 */
 			$resource_instance = $reflectionClass->newInstanceArgs([
-				$this->getHttpClient(),
+				$this->httpClient,
 				$this->client_id,
 				$this->client_secret,
 				$this->plaidEnvironments[$this->environment]
@@ -130,30 +104,5 @@ class Plaid
 		}
 
 		return $this->resource_cache[$resource];
-	}
-
-	/**
-	 * Set a specific ClientInterface instance to be used to make HTTP calls.
-	 *
-	 * @param ClientInterface $httpClient
-	 * @return void
-	 */
-	public function setHttpClient(ClientInterface $httpClient): void
-	{
-		$this->httpClient = $httpClient;
-	}
-
-	/**
-	 * Get the ClientInterface instance being used to make HTTP calls.
-	 *
-	 * @return ClientInterface
-	 */
-	public function getHttpClient(): ClientInterface
-	{
-		if( empty($this->httpClient) ){
-			$this->httpClient = new Shuttle;
-		}
-
-		return $this->httpClient;
 	}
 }
