@@ -2,66 +2,24 @@
 
 namespace TomorrowIdeas\Plaid\Entities;
 
-class RecipientAddress
+use JsonSerializable;
+
+class RecipientAddress implements JsonSerializable
 {
 	/**
-	 * Street address.
-	 *
-	 * @var string
-	 */
-	protected $street;
-
-	/**
-	 * Additional street address.
-	 *
-	 * @var string|null
-	 */
-	protected $street2;
-
-	/**
-	 * City
-	 *
-	 * @var string
-	 */
-	protected $city;
-
-	/**
-	 * Postal code
-	 *
-	 * @var string
-	 */
-	protected $postal_code;
-
-	/**
-	 * Country (2 character ISO)
-	 *
-	 * @var string
-	 */
-	protected $country;
-
-	/**
-	 * Address constructor.
-	 *
-	 * The Address object is needed for certain requests to Plaid.
-	 *
 	 * @param string $street
 	 * @param string|null $street2
 	 * @param string $city
 	 * @param string $postal_code
-	 * @param string $country
+	 * @param string $country Country (2 character ISO)
 	 */
 	public function __construct(
-		string $street,
-		?string $street2,
-		string $city,
-		string $postal_code,
-		string $country)
+		protected string $street,
+		protected ?string $street2,
+		protected string $city,
+		protected string $postal_code,
+		protected string $country)
 	{
-		$this->street = $street;
-		$this->street2 = $street2;
-		$this->city = $city;
-		$this->postal_code = $postal_code;
-		$this->country = $country;
 	}
 
 	/**
@@ -69,7 +27,7 @@ class RecipientAddress
 	 *
 	 * @return array
 	 */
-	public function toArray(): array
+	public function jsonSerialize(): mixed
 	{
 		$streetAddress = [$this->street];
 
